chore(examples): move constructor tests to example constructor
