+++ examples/constructor.rs
@@ -31,15 +31,44 @@ pub fn constructor(value: i32) {
 #[cfg(not(target_arch = "wasm32"))]
 fn main() {}
 
-#[test]
-fn test() -> anyhow::Result<()> {
-    use zint::Contract;
+#[cfg(test)]
+mod tests {
+    use zint::{
+        ethers::abi::{Abi, Function, Param, ParamType, StateMutability},
+        Contract, Ethers,
+    };
 
-    let _contract = Contract::search("constructor")?
-        .constructor(true)
-        .compile()?;
+    #[tokio::test]
+    #[allow(deprecated)]
+    async fn constructor() -> zint::Result<()> {
+        let api = Ethers::anvil()?;
+        let bytecode = Contract::search("constructor")?
+            .constructor(true)
+            .compile()?
+            .bytecode;
 
-    // let info = contract.execute(["get()"])?;
-    // assert_eq!(info.ret, 1.to_bytes32());
-    Ok(())
+        // TODO: Generate ABI issue #47
+        let mut abi: Abi = Default::default();
+        abi.functions.insert(
+            "get".into(),
+            vec![Function {
+                name: "get".into(),
+                inputs: Default::default(),
+                outputs: vec![Param {
+                    name: Default::default(),
+                    kind: ParamType::Int(32usize),
+                    internal_type: None,
+                }],
+                constant: None,
+                state_mutability: StateMutability::View,
+            }],
+        );
+
+        let factory = api.factory(abi, bytecode)?;
+        let contract = factory.deploy(())?.legacy().send().await?;
+        let r = contract.method::<(), i32>("get", ())?.call().await?;
+
+        assert_eq!(r, 1);
+        Ok(())
+    }
 }
